FFS-2494: Implement /summary and /submit pages (#477)
Previously, the income data was shown on the /summary page and included a legal consent checkbox.
Now, the consent checkbox has been moved to the new /submit page.
Additionally, the pdf generation code was moved under the /submit page to keep things Rails-y.
