+++ app/app/components/pinwheel_data_point_component.rb
@@ -18,7 +18,7 @@ class PinwheelDataPointComponent < ViewComponent::Base
   def pay_period_with_frequency(start_date, end_date, pay_frequency)
     translated_pay_frequency = translate_pinwheel_value("payment_frequencies", pay_frequency)
     {
-      label: I18n.t("cbv.summaries.show.pdf.caseworker.pay_period", pay_frequency: translated_pay_frequency),
+      label: I18n.t("cbv.submits.show.pdf.caseworker.pay_period", pay_frequency: translated_pay_frequency),
       value: I18n.t("cbv.payment_details.show.pay_period_value", start_date: format_date(start_date), end_date: format_date(end_date))
     }
   end
@@ -115,14 +115,14 @@ class PinwheelDataPointComponent < ViewComponent::Base
 
   def employer_address(address)
     {
-      label: I18n.t("cbv.summaries.show.pdf.client.address"),
+      label: I18n.t("cbv.submits.show.pdf.client.address"),
       value: address
     }
   end
 
   def client_full_name(full_name)
     {
-      label: I18n.t("cbv.summaries.show.pdf.caseworker.client_full_name"),
+      label: I18n.t("cbv.submits.show.pdf.caseworker.client_full_name"),
       value: full_name
     }
   end
