+++ app/app/mailers/caseworker_mailer.rb
@@ -33,7 +33,7 @@ class CaseworkerMailer < ApplicationMailer
     I18n.with_locale(:en) do
       # caseworkers should receive the report in English
       WickedPdf.new.pdf_from_string(
-        render_to_string(template: "cbv/summaries/show", layout: "pdf", formats: [ :pdf ], locals: { is_caseworker: true }),
+        render_to_string(template: "cbv/submits/show", layout: "pdf", formats: [ :pdf ], locals: { is_caseworker: true }),
         footer: { right: "Income Verification Report | Page [page] of [topage]", font_size: 10 }
       )
     end
