+++ app/app/views/cbv/submits/show.html.erb
@@ -1,0 +1,17 @@
+<% title = t(".page_header") %>
+<% content_for :title, title %>
+<div class="grid-col tablet:grid-col-10">
+<h1><%= title %></h1>
+</div>
+
+<div class="usa-prose">
+  <p><%= t(".warning_html") %></p>
+</div>
+
+<h2><%= t(".legal_header") %></h2>
+<div class="usa-prose">
+  <%= form_with model: @cbv_flow, url: cbv_flow_submit_path, html: { class: "usa-form maxw-full" }, builder: UswdsFormBuilder, method: "patch" do |f| %>
+  <%= f.check_box :consent_to_authorized_use, { 'label': t(".consent_checkbox_label_html", agency_acronym: current_agency.agency_short_name) } %>
+  <%= f.submit t(".share_report_button", agency_acronym: current_agency.agency_short_name), class: "margin-top-5" %>
+  <% end %>
+</div>
