+++ app/app/views/cbv/successes/show.html.erb
@@ -12,13 +12,13 @@
 
   <p><%= agency_translation(".check_status") %></p>
 
-<%= link_to cbv_flow_summary_path(format: "pdf"), target: "_blank", rel: "noopener noreferrer", class: "usa-button margin-top-3 usa-button--primary" do %>
-  <%= t(".download") %>
-    <% end %>
+  <%= link_to cbv_flow_submit_path(format: "pdf"), target: "_blank", rel: "noopener noreferrer", class: "usa-button margin-top-3 usa-button--primary" do %>
+    <%= t(".download") %>
+      <% end %>
 
-<div>
-  <%= link_to agency_url, target: "_blank", rel: "noopener noreferrer", class: "usa-button margin-top-3 usa-button usa-button--outline" do %>
-    <%= t(".back_to_agency", agency_short_name: current_agency.agency_short_name) %>
-  <% end %>
-</div>
+  <div>
+    <%= link_to agency_url, target: "_blank", rel: "noopener noreferrer", class: "usa-button margin-top-3 usa-button usa-button--outline" do %>
+      <%= t(".back_to_agency", agency_short_name: current_agency.agency_short_name) %>
+    <% end %>
+  </div>
 </div>
