+++ app/config/i18n-tasks.yml
@@ -132,9 +132,10 @@ ignore_missing:
     - "activerecord.errors.models.cbv_applicant*"
     - "caseworker_mailer.summary_email.*"
     # Caseworker PDF strings that are only supported in English:
-    - "cbv.summaries.show.pdf.caseworker.*"
+    - "cbv.submits.show.pdf.caseworker.*"
     # Pinwheel values that are only supported in English:
     - "pinwheel.*"
+    # Currently awaiting translation:
     - "cbv.expired_invitations.show.body_2"
     - "help.*"
     - "shared.header.help"
