+++ app/config/routes.rb
@@ -31,7 +31,8 @@ Rails.application.routes.draw do
       resource :employer_search, only: %i[show]
       resource :synchronizations, only: %i[show update]
       resource :synchronization_failures, only: %i[show]
-      resource :summary, only: %i[show update], format: %i[html pdf]
+      resource :summary, only: %i[show update]
+      resource :submit, only: %i[show update], format: %i[html pdf]
       resource :missing_results, only: %i[show]
       resource :success, only: %i[show]
       resource :add_job, only: %i[show create]
